+++ CHANGELOG.md
@@ -13,6 +13,9 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Fixed
 - [Patch] Fix tiny README mistakes. (#451)
 
+### Fixed
+- [Patch] Add selector for `oui-button` in `_button-row.scss`. (#455)
+
 ## [13.0.1][13.0.1] - 2016-07-13
 ### Added
 - [Patch] Add Google Analytics to React documentation. (#448)